# specs/003-resolve-critical-issues/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GASTAT International Dossier System API
  description: API for managing international relations, documents, and intelligence data
  version: 1.0.0
  contact:
    name: GASTAT Development Team
    email: dev@gastat.gov.sa
servers:
  - url: https://api.gastat.gov.sa/v1
    description: Production server
  - url: https://staging-api.gastat.gov.sa/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /countries:
    get:
      summary: List countries
      description: Get paginated list of countries with filtering
      tags: [Countries]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
            example: "Saudi Arabia"
        - name: region
          in: query
          schema:
            type: string
            enum: [asia, africa, europe, americas, oceania]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
        - name: iso_code
          in: query
          schema:
            type: string
            pattern: '^[A-Z]{2,3}$'
            example: "SA"
      responses:
        '200':
          description: List of countries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Country'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  filters:
                    $ref: '#/components/schemas/CountryFilters'
        '400':
          description: Invalid query parameters

    post:
      summary: Create country
      description: Create a new country
      tags: [Countries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCountryRequest'
      responses:
        '201':
          description: Country created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        '400':
          description: Invalid input data
        '409':
          description: Country already exists
  /countries/{id}:
    get:
      summary: Get country by ID
      description: Retrieve a specific country by ID
      tags: [Countries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Country details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        '404':
          description: Country not found

    put:
      summary: Update country
      description: Update an existing country
      tags: [Countries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCountryRequest'
      responses:
        '200':
          description: Country updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        '400':
          description: Invalid input data
        '404':
          description: Country not found

    delete:
      summary: Delete country
      description: Soft delete a country
      tags: [Countries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Country deleted successfully
        '404':
          description: Country not found

  # Organizations Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Country:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name_en:
          type: string
        name_ar:
          type: string
        iso_alpha2:
          type: string
          pattern: '^[A-Z]{2}$'
        iso_alpha3:
          type: string
          pattern: '^[A-Z]{3}$'
        region:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCountryRequest:
      type: object
      required: [name_en, name_ar, iso_alpha2, iso_alpha3, region]
      properties:
        name_en:
          type: string
          maxLength: 255
        name_ar:
          type: string
          maxLength: 255
        iso_alpha2:
          type: string
          pattern: '^[A-Z]{2}$'
        iso_alpha3:
          type: string
          pattern: '^[A-Z]{3}$'
        region:
          type: string
          enum: [asia, africa, europe, americas, oceania]
        status:
          type: string
          enum: [active, inactive, suspended]
          default: active
    UpdateCountryRequest:
      type: object
      properties:
        name_en:
          type: string
          maxLength: 255
        name_ar:
          type: string
          maxLength: 255
        region:
          type: string
          enum: [asia, africa, europe, americas, oceania]
        status:
          type: string
          enum: [active, inactive, suspended]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
    CountryFilters:
      type: object
      properties:
        regions:
          type: array
          items:
            type: string
        statuses:
          type: array
          items:
            type: string
        iso_codes:
          type: array
          items:
            type: string
...